fixup! chore(logging): add logging provider; refactor settings to use logging service from context
