+++ packages/compass-home/src/index.ts
@@ -1,5 +1,5 @@
 import { registerHadronPlugin } from 'hadron-app-registry';
-import Home from './plugin';
+import Home from './components/home';
 
 /**
  * Activate all the components in the Home package.
@@ -18,16 +18,13 @@ function deactivate(): void {
 export const CompassHomePlugin = registerHadronPlugin({
   name: 'CompassHome',
   component: Home,
-  activate(_, { localAppRegistry }) {
+  activate(/* ..., { globalAppRegistry, localAppRegistry } */) {
     // TODO: This is where we should be subscribing to appRegistry events
     // instead of passing it directly to the Home component. Keeping it as-is
     // as cleaning up compass-home is a bigger refactor that is not in scope
     return {
-      // Not the purpose of this interface, just for compatibility we make it
-      // work in a way that will not break compass-home dependency on app
-      // registry that should be removed
       store: {
-        state: { localAppRegistry },
+        state: {},
       },
       deactivate() {
         /* noop */
