+++ packages/compass-import-export/src/components/import-toast.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Body, css, openToast } from '@mongodb-js/compass-components';
 import path from 'path';
-import createLoggerAndTelemetry from '@mongodb-js/compass-logging';
+import { createLoggerAndTelemetry } from '@mongodb-js/compass-logging';
 
 import { ToastBody } from './toast-body';
 import { openFile } from '../utils/open-file';
