+++ packages/compass-logging/src/index.ts
@@ -1,5 +1,4 @@
-export { createLoggerAndTelemetry } from './logger';
-export { createLoggerAndTelemetry as default } from './logger';
+export { createLoggerAndTelemetry } from './ipc-logger';
 export type { LoggerAndTelemetry } from './logger';
 export { mongoLogId } from 'mongodb-log-writer';
 import createDebug from 'debug';
