+++ packages/compass-logging/src/logger.ts
@@ -1,9 +1,7 @@
 import type { MongoLogEntry } from 'mongodb-log-writer';
 import { MongoLogWriter, mongoLogId } from 'mongodb-log-writer';
-import isElectronRenderer from 'is-electron-renderer';
 import createDebug from 'debug';
 import type { Writable } from 'stream';
-import type { HadronIpcRenderer } from 'hadron-ipc';
 
 let preferences: {
   getPreferences(): { trackUsageStatistics: boolean };
@@ -12,19 +10,6 @@ let preferences: {
 type TrackProps = Record<string, any> | (() => Record<string, any>);
 type TrackFunction = (event: string, properties?: TrackProps) => void;
 
-function emit(
-  ipc: HadronIpcRenderer | null | undefined,
-  event: string,
-  data: Record<string, any>
-): void {
-  // We use ipc.callQuiet instead of ipc.call because we already
-  // print debugging messages below
-  void ipc?.callQuiet?.(event, data);
-  if (typeof process !== 'undefined' && typeof process.emit === 'function') {
-    (process as any).emit(event, data);
-  }
-}
-
 export type LoggerAndTelemetry = {
   log: ReturnType<MongoLogWriter['bindComponent']>;
   mongoLogId: typeof mongoLogId;
@@ -32,22 +17,17 @@ export type LoggerAndTelemetry = {
   track: TrackFunction;
 };
 
-export function createLoggerAndTelemetry(
-  component: string
+export function createGenericLoggerAndTelemetry(
+  component: string,
+  emit: (event: string, arg: any) => void
 ): LoggerAndTelemetry {
-  // This application may not be running in an Node.js/Electron context.
-  const ipc: HadronIpcRenderer | null | undefined = isElectronRenderer
-    ? // eslint-disable-next-line @typescript-eslint/no-var-requires
-      require('hadron-ipc').ipcRenderer
-    : null;
-
   // Do not create an actual Writable stream here, since the callback
   // logic in Node.js streams would mean that two writes from the
   // same event loop tick would not be written synchronously,
   // allowing another logger's write to be written out-of-order.
   const target = {
     write(line: string, callback: () => void) {
-      emit(ipc, 'compass:log', { line });
+      emit('compass:log', { line });
       callback();
     },
     end(callback: () => void) {
@@ -99,7 +79,7 @@ export function createLoggerAndTelemetry(
         // for instance if we can't fetch host information,
         // we track a new error indicating the failure.
         // This is so that we are aware of which events might be misreported.
-        emit(ipc, 'compass:track', {
+        emit('compass:track', {
           event: 'Error Fetching Attributes',
           properties: {
             event_name: event,
@@ -118,7 +98,7 @@ export function createLoggerAndTelemetry(
         return;
       }
     }
-    emit(ipc, 'compass:track', data);
+    emit('compass:track', data);
   };
 
   const debug = createDebug(`mongodb-compass:${component.toLowerCase()}`);
@@ -136,5 +116,3 @@ export function createLoggerAndTelemetry(
     track,
   };
 }
-
-export default createLoggerAndTelemetry;
