+++ packages/compass-query-bar/src/stores/ai-query-reducer.ts
@@ -1,5 +1,5 @@
 import type { Reducer } from 'redux';
-import createLoggerAndTelemetry from '@mongodb-js/compass-logging';
+import { createLoggerAndTelemetry } from '@mongodb-js/compass-logging';
 import { getSimplifiedSchema } from 'mongodb-schema';
 import toNS from 'mongodb-ns';
 import preferences from 'compass-preferences-model';
