+++ packages/compass-shell/src/stores/store.js
@@ -3,7 +3,7 @@ import reducer from '../modules';
 import { changeEnableShell, setupRuntime } from '../modules/runtime';
 import { globalAppRegistryActivated } from '@mongodb-js/mongodb-redux-common/app-registry';
 import { setupLoggerAndTelemetry } from '@mongosh/logging';
-import createLoggerAndTelemetry from '@mongodb-js/compass-logging';
+import { createLoggerAndTelemetry } from '@mongodb-js/compass-logging';
 import preferences from 'compass-preferences-model';
 
 const { log, debug, track } = createLoggerAndTelemetry('COMPASS-SHELL');
